+++ PrimeC/solution_5/run.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 # Serial versions
-for x in sieve_memcopy2 primes_words; do
+for x in sieve_memcopy; do
     ./$x
 done
 # Thread parallel versions, computing each sieve quicker.
