+++ PrimeC/solution_5/test.sh
@@ -1,6 +1,4 @@
 #!/bin/sh
 CC="gcc -Ofast -march=native -mtune=native -funroll-all-loops" 
-for x in $1; do
-    $CC -o $x $x.c -lm
-done
-./$1
+$CC -o $1 $1.c -lm
+./$1 $2 $3 $4 $5 $6 $7
